+++ src/Armor.cpp
@@ -1,7 +1,7 @@
 // Armor.cpp
 #include "Armor.h"
 
-Armor::Armor(const std::string& name, int def, int hp)
+Armor::Armor(const std::string& name, double def, double hp)
     : Item(name, ItemType::ARMOR, def), bonusHp(hp) {}
 
 void Armor::applyEffect(Character& character) {
