+++ src/Item.cpp
@@ -1,9 +1,9 @@
 // Item.cpp
 #include "Item.h"
 
-Item::Item(const std::string& name, ItemType type, int value)
+Item::Item(const std::string& name, ItemType type, double value)
     : name(name), type(type), value(value) {}
 
 const std::string& Item::getName() const { return name; }
 ItemType Item::getType() const { return type; }
-int Item::getValue() const { return value; }
+double Item::getValue() const { return value; } // int -> double
