+++ src/Item.h
@@ -5,24 +5,24 @@
 #include <string>
 #include "Character.h"
 
-enum class ItemType { WEAPON, ARMOR, POTION, ACCESSORY, AMULET };
+enum class ItemType { WEAPON, ARMOR, POTION, ACCESSORY, AMULET }; // agregue el quinto item que faltaba definir
 
 class Item {
 public:
-    Item(const std::string& name, ItemType type, int value);
-    virtual ~Item() = default;
+    Item(const std::string& name, ItemType type, double value);
+    virtual ~Item() = default; // para la sobrescritura de las subclases
 
     const std::string& getName() const;
     ItemType getType() const;
-    int getValue() const;
+    double getValue() const; // int -> double
 
-    virtual void applyEffect(Character& character) = 0;
+    virtual void applyEffect(Character& character) = 0; // aplica el efecto de los items al personaje
 
 
 private:
     std::string name;
     ItemType type;
-    int value;
+    double value; // ya que los atributos del personaje son double, cambie el tipo de Value por double tambien
 };
 
 #endif // ITEM_H
