+++ src/Weapon.cpp
@@ -1,7 +1,7 @@
 // Weapon.cpp
 #include "Weapon.h"
 
-Weapon::Weapon(const std::string& name, int atk, int lck)
+Weapon::Weapon(const std::string& name, double atk, double lck)
     : Item(name, ItemType::WEAPON, atk), bonusLck(lck) {}
 
 void Weapon::applyEffect(Character& character) {
